brute force method for reducing ADC noise -- average a ridiculous number of samples  (because, for some reason, even though 64 samples is plenty in a test program,   it ends up being extremely erratic when used inside Anduril...  and I'm not sure why) also, use 15-bit ADC values instead of 16 bits, in the temperature logic  (to help protect against integer overflows) ... but this code still doesn't work well.  It regulates down *very* fast, and then   gradually rises until the next extra-fast drop-down.  :( ... also, tempcheck mode sometimes changes by 4-5 C between readouts, which is worrisome. ... and factory reset is still broken.

+++ spaghetti-monster/fsm-adc.h
@@ -40,7 +40,7 @@
 #endif
 
 volatile uint8_t irq_adc = 0;  // ADC interrupt happened?
-uint8_t adc_sample_count = 0;  // skip the first sample; it's junk
+uint16_t adc_sample_count = 0;  // skip the first sample; it's junk
 uint8_t adc_channel = 0;  // 0=voltage, 1=temperature
 uint16_t adc_raw[2];  // last ADC measurements (0=voltage, 1=temperature)
 uint16_t adc_smooth[2];  // lowpassed ADC measurements (0=voltage, 1=temperature)
